+++ jni/mesh.h
@@ -184,16 +184,16 @@ public:
      * Down event method. Called when entity is pressed down
      * on touchscreen.
      */
-    virtual void down(int x, int y){LOGE("Called generic down method");};
+    virtual void down(float x, float y){LOGE("Called generic down method");};
     /**
      * Move event method. Called when finger is moved on 
      * touchscreen with this mesh pressed.
      */
-    virtual void move(int x, int y){};
+    virtual void move(float x, float y){};
     /**
      * Up event method. Called when finger is lifted.
      */
-    virtual void up(int x, int y){LOGE("Called generic up method");};
+    virtual void up(float x, float y){LOGE("Called generic up method");};
 private:
     void _setBuffer(GLenum target, GLfloat *buf, GLuint size, GLuint sel);
 };
@@ -254,8 +254,8 @@ public:
     void setStateTexture(int state, GLfloat x, GLfloat y);
     void setTextureSize(GLfloat sx, GLfloat xy);
     
-    void down(int x, int y);
-    void up(int x, int y);
+    void down(float x, float y);
+    void up(float x, float y);
 };
 
 #endif /* MESH_H_ */
